good place to lock in before the input stream code
